requirements.txt

pipecat-ai
baml-py
openai
python-dotenv
asyncio
websockets
requests